+++ contact_book_system/backend/app/api/chat.py
@@ -1,11 +1,11 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 
-from routes.db.db import get_db
-from routes.models.accounts_model import Account, RoleEnum
-from routes.services.auth import get_current_user
-from routes.services.chat_service import ChatService
-from routes.schemas.chat_schema import (
+from app.db.db import get_db
+from app.models.accounts_model import Account, RoleEnum
+from app.services.auth import get_current_user
+from app.services.chat_service import ChatService
+from app.schemas.chat_schema import (
     ChatRoomCreate,
     ChatRoomUpdate,
     ChatRoomResponse,