+++ contact_book_system/backend/app/models/password_reset_model.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String, Boolean, TIMESTAMP, ForeignKey
 from sqlalchemy.sql import func
-from routes.db.base import Base
+from app.db.base import Base
 
 
 class PasswordResetToken(Base):